Restructure Test project
