+++ NetDaemonApps.Test/GlobalUsings.cs
@@ -1,0 +1,5 @@
+﻿global using FluentAssertions;
+global using HassModel;
+global using Moq;
+global using NetDaemon.HassModel;
+global using NetDaemon.HassModel.Entities;
